Fix failure to parse an empty program
This resolves #10 in a more natural way. The final statement in a
statement list can be empty (i.e. there can be a delimiter after the
last true statement and there doesn't have to be). This has no effect on
interpreting the AST since in SQF the last returned value is returned
(i.e. a final NoOp doesn't mean the return is nil).