Introduce preprocessing step
Decided to introduce a new class for better functional encapsulation and
clear distribution of duties. First step is to handle comments in the
preprocessor since that's where they are actually stripped out which
results in some slightly different behaviour of tokenisation (added
tests to capture this).

On the road to #13