Fix lookahead logic in preprocessor
Acting as a wrapper for the stream peek meant if the next character was
a comment then the lexer could produce two seperate tokens in cases
where it should produce only one.

Introduced a buffer to reuse the existing `get` logic for peeking and
thus preserve preprocessing easily. Will also come in handy when it
comes time to perform macro replacement.