fix(pp): prevent recursive file inclusion
This was both problematic because it created an infinite loop and
because the RV engine crashes when this is encountered, which should be
treated as an error.

The preprocessor class is quite bloated and should probably be split
into components with narrower responsibilities in future. The
constructors could certainly be cleaned up with delegation too.