feat: group errors by file in output
To achieve this errors are stored in a map structure (so that output is
sorted) with unique pointers. After analysis of each file they are then
printed out in an appropriate format by the main program.

This also fixes the failing analysis and rules tests which can now more
specifically check the details of errors encountered instead of being at
the whim of the output formatting.

closes #33