fix: prevent empty filename for some tokens
When words were encountered in the preprocessor, if they turned out to
be a regular id then the characters were constructed with properties
cached from the encountered initial. Since storing the filename was
added later I forgot to update this point in the code, I now copy all
properties of the initial to prevent that happening again.