feat(pp): introduce source consumer component
Breaking the preprocessor down into parts with individual
responsibilities. This part purely retrieves characters out of a source
stream and assigns them a position.

The intention is that this component can be passed to other parts of the
preprocessor to enable them all to consume the source stream in a shared
fashion.